+++ src/lib.rs
@@ -39,20 +39,23 @@
 #[cfg(test)]
 extern crate std;
 
-const D: usize = 20;
+use core::{
+    cmp::PartialEq,
+    convert::TryFrom,
+    ops::{Add, AddAssign, Div},
+};
 
 /// Returns the number at `index` of the Halton sequence for `base`. The number
-/// returned will be > 0 and < 1, assuming index > 1.
+/// returned will be > 0 and < 1, assuming `index` > 1.
 ///
-/// While [`Sequence`](struct.Sequence.html) will be faster for most cases,
-/// this function may be useful for calulating a single number from a Halton
-/// sequence, or creating a 'leaped' sequence.
+/// While [`Sequence`] will be faster for most cases, this function may be
+/// useful for calulating a single number from a Halton sequence, or creating
+/// a 'leaped' sequence.
 ///
 /// # Index
 ///
-/// Beware that indexing [`Sequence`](struct.Sequence.html) is effectively
-/// 0-based, whereas the `index` argument for [`number`](fn.number.html) is
-/// 1-based.
+/// Beware that indexing [`Sequence`] is effectively 0-based, whereas the
+/// `index` argument for [`number`] is 1-based.
 ///
 /// ```
 /// use halton::{number, Sequence};
@@ -73,7 +76,7 @@ const D: usize = 20;
 /// }
 /// ```
 #[inline]
-pub fn number(base: u8, mut index: usize) -> f64 {
+pub fn number(base: u16, mut index: usize) -> f64 {
     let mut factor = 1.0;
     let mut result = 0.0;
     while index > 0 {
@@ -86,12 +89,15 @@ pub fn number(base: u8, mut index: usize) -> f64 {
 
 /// An iterator implementing the fast generation of Halton sequences.
 /// The method of generation is adapted from _Fast, portable, and reliable
-/// algorithm for the calculation of Halton numbers_ by Miroslav Kolar and
-/// Seamus O'Shea.
+/// algorithm for the calculation of Halton numbers_ by Miroslav Kolář and
+/// Seamus F. O'Shea.
 ///
 /// The numbers returned from the iterator will be in the range > 0 and < 1.
 ///
 /// # Examples
+///
+/// These examples use [`Sequence`], this is an alias for [`GenericSequence`]
+/// with default type parameters.
 ///
 /// With a `for` loop:
 ///
@@ -126,31 +132,37 @@ pub fn number(base: u8, mut index: usize) -> f64 {
 /// assert_eq!(Some(0.23875432525951557), seq.next());
 /// ```
 #[derive(Clone)]
-pub struct Sequence {
-    b: u8,
-    d: [u8; D],
-    r: [f64; D],
+pub struct GenericSequence<I, F, const D: usize> {
+    b: I,
+    d: [I; D],
+    r: [F; D],
 }
 
-impl Sequence {
-    /// Constructs a new `Sequence` for `base`.
+impl<I, F, const D: usize> GenericSequence<I, F, D>
+where
+    I: AddAssign + Copy + From<u8> + PartialEq + TryFrom<usize>,
+    f64: From<I>,
+    usize: From<I>,
+    F: Add<Output = F> + Copy + Div<Output = F> + From<I> + From<f32>,
+{
+    /// Constructs a new [`GenericSequence`] for `base`.
     ///
     /// # Examples
     ///
     /// Basic usage:
     ///
     /// ```
-    /// # use halton::Sequence;
-    /// let mut seq = Sequence::new(2);
+    /// # use halton::GenericSequence;
+    /// let mut seq = GenericSequence::<u16, f64, 20>::new(2);
     ///
     /// assert_eq!(Some(0.5), seq.next());
     /// ```
     #[inline]
-    pub fn new(base: u8) -> Self {
-        Sequence {
+    pub fn new(base: I) -> Self {
+        GenericSequence {
             b: base,
-            d: [0; D],
-            r: [0.0; D],
+            d: [I::from(0u8); D],
+            r: [F::from(0.0f32); D],
         }
     }
 
@@ -158,12 +170,14 @@ impl Sequence {
         self.d
             .iter()
             .zip(1..)
-            .map(|(v, i)| (*v as usize).checked_mul(i))
+            .map(|(v, i)| usize::from(*v).checked_mul(i))
             .try_fold(0usize, |acc, v| acc.checked_add(v?))
     }
 
     fn max(&self) -> Option<usize> {
-        _checked_pow(self.b as usize, D).map(|v| v - 1)
+        u32::try_from(self.d.len())
+            .ok()
+            .and_then(|len| usize::from(self.b).checked_pow(len).map(|v| v - 1))
     }
 
     fn remaining(&self) -> Option<usize> {
@@ -171,30 +185,36 @@ impl Sequence {
     }
 }
 
-impl Iterator for Sequence {
-    type Item = f64;
+impl<I, F, const D: usize> Iterator for GenericSequence<I, F, D>
+where
+    I: AddAssign + Copy + From<u8> + PartialEq + TryFrom<usize>,
+    f64: From<I>,
+    usize: From<I>,
+    F: Add<Output = F> + Copy + Div<Output = F> + From<I> + From<f32>,
+{
+    type Item = F;
 
     #[inline]
     fn next(&mut self) -> Option<Self::Item> {
         let mut l = 0;
 
-        self.d[l] += 1;
+        self.d[l] += I::from(1u8);
         if self.d[l] == self.b {
             while self.d[l] == self.b {
-                self.d[l] = 0;
+                self.d[l] = I::from(0u8);
                 l += 1;
-                if l == D {
+                if l == self.d.len() {
                     return None;
                 }
-                self.d[l] += 1;
+                self.d[l] += I::from(1u8);
             }
-            self.r[l - 1] = (f64::from(self.d[l]) + self.r[l]) / f64::from(self.b);
+            self.r[l - 1] = (F::from(self.d[l]) + self.r[l]) / F::from(self.b);
             for i in (1..l).rev() {
-                self.r[i - 1] = self.r[i] / f64::from(self.b);
+                self.r[i - 1] = self.r[i] / F::from(self.b);
             }
-            Some(self.r[0] / f64::from(self.b))
+            Some(self.r[0] / F::from(self.b))
         } else {
-            Some((f64::from(self.d[0]) + self.r[0]) / f64::from(self.b))
+            Some((F::from(self.d[0]) + self.r[0]) / F::from(self.b))
         }
     }
 
@@ -208,10 +228,7 @@ impl Iterator for Sequence {
     }
 
     #[inline]
-    fn count(self) -> usize
-    where
-        Self: Sized,
-    {
+    fn count(self) -> usize {
         if let Some(remaining) = self.remaining() {
             remaining
         } else {
@@ -220,10 +237,7 @@ impl Iterator for Sequence {
     }
 
     #[inline]
-    fn last(mut self) -> Option<Self::Item>
-    where
-        Self: Sized,
-    {
+    fn last(mut self) -> Option<Self::Item> {
         if let Some(remaining) = self.remaining() {
             self.nth(remaining - 1)
         } else {
@@ -235,19 +249,19 @@ impl Iterator for Sequence {
     fn nth(&mut self, mut n: usize) -> Option<Self::Item> {
         if n > 50 {
             if let Some(mut n) = self.pos().and_then(|p| n.checked_add(p)) {
-                self.d.iter_mut().for_each(|v| *v = 0);
-                self.r.iter_mut().for_each(|v| *v = 0.0);
+                self.d.iter_mut().for_each(|v| *v = I::from(0u8));
+                self.r.iter_mut().for_each(|v| *v = F::from(0.0f32));
                 let mut last = 0;
                 while n >= usize::from(self.b) {
-                    self.d[last] = n as u8 % self.b;
+                    self.d[last] = I::try_from(n % usize::from(self.b)).ok().unwrap();
                     last += 1;
                     n /= usize::from(self.b);
                 }
-                self.d[last] = n as u8;
-                for i in (1..D).rev() {
-                    self.r[i - 1] = (f64::from(self.d[i]) + self.r[i]) / f64::from(self.b);
+                self.d[last] = I::try_from(n).ok().unwrap();
+                for i in (1..self.r.len()).rev() {
+                    self.r[i - 1] = (F::from(self.d[i]) + self.r[i]) / F::from(self.b);
                 }
-                return self.next()
+                return self.next();
             }
         }
         for x in self {
@@ -260,37 +274,8 @@ impl Iterator for Sequence {
     }
 }
 
-impl ExactSizeIterator for Sequence {
-    #[inline]
-    fn len(&self) -> usize {
-        self.remaining().unwrap()
-    }
-}
-
-#[deprecated(since = "0.2.1", note = "checked_pow was made public accidently and will be removed in 0.3")]
-#[doc(hidden)]
-pub fn checked_pow(base: usize, exp: usize) -> Option<usize> {
-    _checked_pow(base, exp)
-}
-
-// copied from rust std, '*' replaced with checked_mul()
-fn _checked_pow(mut base: usize, mut exp: usize) -> Option<usize> {
-    let mut acc = 1usize;
-
-    while exp > 1 {
-        if (exp & 1) == 1 {
-            acc = acc.checked_mul(base)?;
-        }
-        exp /= 2;
-        base = base.checked_mul(base)?;
-    }
-
-    if exp == 1 {
-        acc = acc.checked_mul(base)?;
-    }
-
-    Some(acc)
-}
+/// A type alias for [`GenericSequence`] with defaults.
+pub type Sequence = GenericSequence<u16, f64, 20>;
 
 #[cfg(test)]
 mod tests {